Add VideoPreview component and update typography styles to use Poetsen One font
