+++ src/app/globals.css
@@ -2,6 +2,8 @@
 @tailwind components;
 @tailwind utilities;
 
+@import url('https://fonts.googleapis.com/css2?family=Poetsen+One&display=swap');
+
 :root {
   --background: #ffffff;
   --foreground: #171717;
